market_segmentation_insurance/ml_process/requirements.txt
numpy==2.2.4
pandas==2.2.3
scikit-learn==1.6.1
kmodes==0.12.2
prefect==3.4.8
mlflow==3.1.1
psycopg2-binary==2.9.10